% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_segmented.R
\name{bp_segmented}
\alias{bp_segmented}
\title{PieceWise Regression}
\usage{
bp_segmented(y, bp = NULL)
}
\arguments{
\item{y}{Numeric vector}

\item{bp}{If provide, changing point detected will not applied.}

\item{year.origin}{first year of y}

\item{predict}{If true, returned predicted data of piecewise regression.}
}
\description{
PieceWise Regression
}
